OGM-1229 For now, we need to duplicate a lot of methods in our Session class
It was already the case but as Session is also an EntityManager now, we
have more methods in it (and the code in ORM has more private/protected
methods we need to copy).

Note that the methods were deleted in a previous commit to know exactly
what was required.

They will be removed in a subsequent commit relying on the new
CustomLoaderFactory mechanism introduced in ORM.